+++ src/main.rs
@@ -56,7 +56,7 @@ fn main() -> Result<()> {
 
             println!();
             println!("Mouse actions:");
-            println!(" - scroll(-100)");
+            println!(" - wheel(-100)");
             println!(" - click(left+right)");
             println!(" - move(5,0)");
             println!(" - drag(left+right,0,5)");